@@ -4,18 +4,23 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "build": "uglifyjs lib/ajaxhook.core.min.js Mix_Combine.js lib/min.js -o tmp.js &&  npm run CatHeader && rm -f tmp.js",
-    "CatHeader": "rm -f dist/dist.js && (cat header.js tmp.js >> dist/dist.js)"
+    "CatHeader": "rm -f dist/dist.js && (cat header.js dist/dist.combine.js >> dist/dist.js)",
+    "build": "webpack && npm run CatHeader"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "babel-plugin-add-header-comment": "^1.0.3",
-    "uglify-js": "^3.14.2",
-    "uglifyjs": "^2.4.11",
-    "webpack": "^5.53.0",
-    "webpack-cli": "^4.8.0",
-    "webpack-dev-server": "^4.2.1"
+    "aes-decrypter": "3.1.2",
+    "ajax-hook": "2.0.3",
+    "@vue/runtime-dom": "3.2.31",
+    "hls.js": "1.1.5",
+    "parse-url-query-params": "0.0.6",
+    "vue": "3.2.31",
+    "webpack": "5.70.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-uglify-js-plugin": "1.1.9",
+    "ts-loader": "^9.2.8",
+    "typescript": "^4.6.2"
   }
 }
